mnist-flask-app/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
torch==1.6.0
torchvision==0.7.0
Pillow==7.2.0
numpy==1.19.2
